microBenchMark/18_SubPixel图像超分辨率/requirements.txt
/app/paddlepaddle-2.3.2-cp38-cp38-linux_x86_64.whl
ipython==8.6.0
matplotlib==3.4.1
numpy==1.20.2
paddle==1.0.2
paddlepaddle==2.3.2
Pillow==9.3.0
